requirements.txt
opencv-contrib-python
matplotlib
natsort
numpy
scikit-learn
pillow
tensorflow
